+++ source/engine/4chan/thread/parseThreadsResponse.test.js
@@ -11,9 +11,12 @@ describe('4chan.org', () => {
 		expectToEqual(
 			new Chan(FourChanConfig, {
 				messages: {
-					deletedComment: 'Удалённое сообщение',
-					hiddenComment: 'Скрытое сообщение',
-					quotedComment: 'Сообщение'
+					comment: {
+						deleted: 'Deleted comment',
+						hidden: 'Hidden comment',
+						external: 'Comment from another thread',
+						default: 'Comment'
+					}
 				}
 			}).parseThreads(API_RESPONSE, {
 				boardId: 'a'